fix: env config

+++ ubuntu/base.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 # init symbolic links
 for f in 'zshrc' 'tmux.conf' 'env' 'alias' 'git-commit-template'
 do
@@ -17,9 +19,13 @@ sudo apt install git curl exuberant-ctags cmake \
     python3-dev python3-pip xclip zsh tmux httpie flameshot -y
 
 # pip symbolic links
-[ ! -f /usr/bin/pip ] && \
+if [ ! -f /usr/bin/pip ]
+then
     sudo ln -s /usr/bin/pip3 /usr/bin/pip
+fi
 
 # oh my zsh
-[ ! -d $HOME/.oh-my-zsh ] && \
+if [ ! -d $HOME/.oh-my-zsh ]
+then
     git clone git://github.com/robbyrussell/oh-my-zsh.git $HOME/.oh-my-zsh
+fi
